// src/api/generated/models/visitStatusFilter.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BFF Service
 * Backend For Frontend service aggregating microservices
 * OpenAPI spec version: 1.0.0
 */

/**
 * Temporal status filter for visits.
 */
export type VisitStatusFilter =
  (typeof VisitStatusFilter)[keyof typeof VisitStatusFilter];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VisitStatusFilter = {
  today: "today",
  past: "past",
  future: "future",
} as const;